requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
django==4.2.7; python_version >= '3.8'
django-ipware==6.0.0; python_version >= '3.8'
idna==3.5; python_version >= '3.5'
mysqlclient==2.2.0; python_version >= '3.8'
pillow==10.1.0; python_version >= '3.8'
pypng==0.20220715.0
python-ipware==2.0.0; python_version >= '3.7'
qrcode[pil]==7.4.2; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
uwsgi==2.0.23
